Restart into a better style
I restarted the development. This time however I have made it a bit more modular with the ability to implement drivers and with PCI I intend to load the driver if they have the device. Interrupts are however a bit finicky and input must only happen after interrupts are loaded or they just wont work.